# ita_root/ita_api_organization/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Exastro IT Automation API
  description: Exastro IT Automation API
  version: "2.4"
servers:
- url: "https://{host}"
  description: API endpoint
  variables:
    host:
      default: localhost
- url: "http://{host}"
  description: API endpoint
  variables:
    host:
      default: localhost
- url: "{protocol}://{host}:{port}"
  description: API endpoint (Development)
  variables:
    protocol:
      default: https
      enum:
      - http
      - https
    host:
      default: localhost
    port:
      default: "30080"
...
paths:
...
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/maintenance/all/:
    post:
      tags:
      - Menu MaintenanceAll
      description: レコードを一括で登録/更新/廃止/復活する
      operationId: maintenance_all
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              description: 登録/更新/廃止/復活するidとレコードの値を設定
              items:
                $ref: '#/components/schemas/MAINTENANCE_PARAMETERS'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MAINTENANCE_ALL_FORM'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.menu_maintenance_all_controller
...
components:
  schemas:
...
    ERROR_TEMPLATE:
      type: object
      properties:
        result:
          type: string
          example: 001-002
        message:
          type: string
...
    MAINTENANCE_PARAMETERS:
      type: object
      properties:
        parameter:
          $ref: '#/components/schemas/FILTER_PARAMETERS_parameter'
        file:
          $ref: '#/components/schemas/FILTER_PARAMETERS_file'
        type:
          type: string
          example: Register/Update
...
    MAINTENANCE_ALL_FORM:
      type: object
      properties:
        json_parameters:
          type: string
          description: parameter:リクエストとtype:実行種別(登録/更新/廃止/復活)をlist構造で指定。 parameter下では、項目名と値をkey-value形式で指定。
          example: "[{\"rest_key_name\":\"string\",\"type\":\"Register/Update\"}]"
        xxx.rest_key_name:
          type: array
          description: xxx:parametersのlist番号を指定。 rest_key_name:parametersのファイルアップロードの対象を指定。
          items:
            type: string
            format: binary
...
    inline_response_200_6:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: object
        message:
          type: string
      example:
        result: 000-00000
        data: {}
        message: message
...
    FILTER_PARAMETERS_parameter:
      type: object
      properties:
        rest_key_name:
          type: string
      example:
        rest_key_name: rest_key_name
    FILTER_PARAMETERS_file:
      type: object
      properties:
        rest_key_name:
          type: string
          example: base64 encode file
      example:
        rest_key_name: base64 encode file
...
  securitySchemes:
    BearerAuth:
      type: http
      description: Credentials or access token for API
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth
    BasicAuth:
      type: http
      description: Value converted from ID/Password in base64
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_BasicAuth